requirements.txt
opencv-python~=4.11.0.86
moviepy~=2.1.2
python-vlc~=3.0.12117
PySide6~=6.9.0
openai-whisper~=20240930
pillow~=10.4.0
rapidfuzz~=3.13.0